coreiface: deprecate DhtAPI in favor of RoutingAPI
