+++ coreiface/tests/api.go
@@ -75,7 +75,6 @@ func TestApi(p Provider) func(t *testing.T) {
 	return func(t *testing.T) {
 		t.Run("Block", tp.TestBlock)
 		t.Run("Dag", tp.TestDag)
-		t.Run("Dht", tp.TestDht)
 		t.Run("Key", tp.TestKey)
 		t.Run("Name", tp.TestName)
 		t.Run("Object", tp.TestObject)
